Peer Reviewer 1
Good job in using the Masterfile.js to utilize the cross-platform sizing. It took a while for me to figure out that with my code. As for the look of your app itself an error popped up pertaining to the last line of code [module.exports = ApplicationWindow;] that code should be module.exports = MasterView; . Still when I was making the changes the application still didn’t open in my genymotion. It may have been an error on my side with the use of the default master template. That is the only problem I see when I just look through your code I see that its there. Your descriptions as well as your pictures. Keep up the good work in the class.
Best,
Peer Reviewer 1